% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{continuous_toy_data_simulation}
\alias{continuous_toy_data_simulation}
\title{Toy data simulation with continuous treatments}
\usage{
continuous_toy_data_simulation(n)
}
\arguments{
\item{n}{the number of samples to draw from the distribution}
}
\value{
a list containing the covariate matrix, treatment vector, outcome vector, true propensity vector, true marginal outcome vector,
true control potential outcome vector, true treated potential outcome vector, and true treatment effect vector, in that order.
}
\description{
Generates a toy dataset of size \eqn{n} that can be used to experiment with the learners and meta-learners
in this package. The generative process should be easy to learn with linear methods.
}
\examples{
data = continuous_toy_data_simulation(500) # draw a sample
}
