% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvkern.R
\name{cv_klrs}
\alias{cv_klrs}
\title{Kernel ridge regression with Gaussian kernels with cross validation to search for hyper-parameters (implemented with th eKLRS package)}
\usage{
cv_klrs(x, y, weights = NULL, k_folds = NULL, b_range = 10^(seq(-3,
  3, 0.5)), lambda_range = 10^(seq(-3, 3, 0.5)))
}
\arguments{
\item{x}{the input features}

\item{y}{the observed response (real valued)}

\item{weights}{weights for input if doing weighted regression/classification. If set to NULL, no weights are used}

\item{k_folds}{number of folds used in cross validation}

\item{b_range}{the range of Gaussian kernel bandwidths for cross validation}

\item{lambda_range}{the range of ridge regression penalty factor for cross validation}
}
\value{
a list that includes the best b, lambda, best predictions, best model fit
}
\description{
Kernel ridge regression with Gaussian kernels with cross validation to search for hyper-parameters (implemented with th eKLRS package)
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
y = pmax(x[,1], 0) + x[,2] + pmin(x[,3], 0) + rnorm(n)

fit = cv_klrs(x, y)
}

}
