% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_utils.R
\name{predict.learner}
\alias{predict.learner}
\title{Prediction for base learner}
\usage{
\method{predict}{learner}(object, newdata)
}
\arguments{
\item{object}{a learner object}

\item{x}{a matrix of covariates for which to predict a target}
}
\description{
Prediction for base learner
}
