% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_learner.R
\name{predict.R_learner}
\alias{predict.R_learner}
\title{Prediction for R-learner}
\usage{
\method{predict}{R_learner}(object, x)
}
\arguments{
\item{object}{a R-learner object}

\item{x}{a matrix of covariates for which to predict the treatment effect}
}
