% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkern.R
\name{predict.rkern}
\alias{predict.rkern}
\title{predict for rkern}
\usage{
\method{predict}{rkern}(object, newx = NULL, ...)
}
\arguments{
\item{object}{an rkern object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained rkern model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

rkern_fit = rkern(x, w, y)
rkern_est = predict(rkern_fit, x)
}


}
