% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlearner.R
\name{predict.rlearner}
\alias{predict.rlearner}
\title{Prediction for R-learner}
\usage{
\method{predict}{rlearner}(object, x)
}
\arguments{
\item{object}{a R-learner object}

\item{x}{a matrix of covariates for which to predict the treatment effect}
}
\description{
Prediction for R-learner
}
