% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sboost.R
\name{predict.sboost}
\alias{predict.sboost}
\title{predict for sboost}
\usage{
\method{predict}{sboost}(object, newx = NULL, ...)
}
\arguments{
\item{object}{a sboost object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained sboost model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

sboost_fit = sboost(x, w, y)
sboost_est = predict(sboost_fit, x)
}


}
