% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skern.R
\name{predict.skern}
\alias{predict.skern}
\title{predict for skern}
\usage{
\method{predict}{skern}(object, newx = NULL, ...)
}
\arguments{
\item{object}{an skern object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained skern model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

skern_fit = skern(x, w, y)
skern_est = predict(skern_fit, x)
}


}
