% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slasso.R
\name{predict.slasso}
\alias{predict.slasso}
\title{predict for slasso}
\usage{
\method{predict}{slasso}(object, newx = NULL, ...)
}
\arguments{
\item{object}{a slasso object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained slasso model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

slasso_fit = slasso(x, w, y)
slasso_est = predict(slasso_fit, x)
}


}
