% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slearner.R
\name{predict.slearner}
\alias{predict.slearner}
\title{Prediction for U-learner}
\usage{
\method{predict}{slearner}(object, x)
}
\arguments{
\item{object}{a U-learner object}

\item{x}{a matrix of covariates for which to predict the treatment effect}
}
