% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_learner.R
\name{predict.T_learner}
\alias{predict.T_learner}
\title{Prediction for U-learner}
\usage{
\method{predict}{T_learner}(object, x)
}
\arguments{
\item{object}{a U-learner object}

\item{x}{a matrix of covariates for which to predict the treatment effect}
}
\description{
Prediction for U-learner
}
