% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tboost.R
\name{predict.tboost}
\alias{predict.tboost}
\title{predict for tboost}
\usage{
\method{predict}{tboost}(object, newx = NULL, ...)
}
\arguments{
\item{object}{a tboost object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained tboost model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

tboost_fit = tboost(x, w, y)
tboost_est = predict(tboost_fit, x)
}


}
