% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkern.R
\name{predict.tkern}
\alias{predict.tkern}
\title{predict for tkern}
\usage{
\method{predict}{tkern}(object, newx = NULL, ...)
}
\arguments{
\item{object}{an tkern object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained tkern model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

tkern_fit = tkern(x, w, y)
tkern_est = predict(tkern_fit, x)
}


}
