% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_learner.R
\name{predict.U_learner}
\alias{predict.U_learner}
\title{Prediction for U-learner}
\usage{
\method{predict}{U_learner}(object, x)
}
\arguments{
\item{object}{a U-learner object}

\item{x}{a matrix of covariates for which to predict the treatment effect}
}
