% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulasso.R
\name{predict.ulasso}
\alias{predict.ulasso}
\title{predict for ulasso}
\usage{
\method{predict}{ulasso}(object, newx = NULL, s = c("lambda.1se",
  "lambda.min"), ...)
}
\arguments{
\item{object}{a ulasso object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{s}{choose from "lambda.min" or "lambda.1se" for prediction}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained ulasso model
}
\examples{
\dontrun{
n = 100; p = 10

X = matrix(rnorm(n*p), n, p)
W = rbinom(n, 1, 0.5)
Y = pmax(X[,1], 0) * W + X[,2] + pmin(X[,3], 0) + rnorm(n)

ulasso.fit = ulasso(X, Y, W)
ulasso.est = predict(ulasso.fit, X)
}


}
