% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulearner.R
\name{predict.ulearner}
\alias{predict.ulearner}
\title{Prediction for U-learner}
\usage{
\method{predict}{ulearner}(object, x)
}
\arguments{
\item{object}{a U-learner object}

\item{x}{a matrix of covariates for which to predict the treatment effect}
}
