% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_learner.R
\name{predict.X_learner}
\alias{predict.X_learner}
\title{Prediction for X-learner}
\usage{
\method{predict}{X_learner}(object, x)
}
\arguments{
\item{object}{a X-learner object}

\item{x}{a matrix of covariates for which to predict the treatment effect}
}
