% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xboost.R
\name{predict.xboost}
\alias{predict.xboost}
\title{predict for xboost}
\usage{
\method{predict}{xboost}(object, newx = NULL, ...)
}
\arguments{
\item{object}{a xboost object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained xboost model
}
\examples{
\dontrun{
n = 100; p = 10

X = matrix(rnorm(n*p), n, p)
W = rbinom(n, 1, 0.5)
Y = pmax(X[,1], 0) * W + X[,2] + pmin(X[,3], 0) + rnorm(n)

xboost.fit = xboost(X, Y, W)
xboost.est = predict(xboost.fit, X)
}


}
