% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlasso.R
\name{predict.xlasso}
\alias{predict.xlasso}
\title{predict for xlasso}
\usage{
\method{predict}{xlasso}(object, newx = NULL, s = c("lambda.min",
  "lambda.1se"), ...)
}
\arguments{
\item{object}{a xlasso object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained xlasso model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

xlasso_fit = xlasso(x, w, y)
xlasso_est = predict(xlasso_fit, x)
}


}
