% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlearner.R
\name{predict.xlearner}
\alias{predict.xlearner}
\title{Prediction for X-learner}
\usage{
\method{predict}{xlearner}(object, x)
}
\arguments{
\item{object}{a X-learner object}

\item{x}{a matrix of covariates for which to predict the treatment effect}
}
\description{
Prediction for X-learner
}
