% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkern.R
\name{rkern}
\alias{rkern}
\title{R-learner, implemented via kernel ridge regression with a Gaussian kernel}
\usage{
rkern(x, w, y, k_folds = NULL, p_hat = NULL, m_hat = NULL,
  b_range = 10^(seq(-3, 3, 0.5)), lambda_range = 10^(seq(-3, 3, 0.5)))
}
\arguments{
\item{x}{the input features}

\item{w}{the treatment variable (0 or 1)}

\item{y}{the observed response (real valued)}

\item{k_folds}{number of folds for cross-fitting}

\item{p_hat}{user-supplied estimate for E[W|X]}

\item{m_hat}{user-supplied estimte for E[Y|X]}

\item{b_range}{the range of Gaussian kernel bandwidths for cross validation}

\item{lambda_range}{the range of ridge regression penalty factor for cross validation}
}
\value{
an rkern object
}
\description{
R-learner, as proposed by Nie and Wager (2017), implemented via kernel ridge regression with a Gaussian kernel
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

rkern_fit = rkern(x, w, y)
rkern_est = predict(rkern_fit, x)
}
}
