% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{rlasso}
\alias{rlasso}
\title{R-learner, as proposed by Nie and Wager, 2017}
\usage{
rlasso(X, Y, W, alpha = 1, nfolds = NULL, lambda.choice = c("lambda.min",
  "lambda.1se"), rs = FALSE, w.hat = NULL, y.hat = NULL)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{W}{the treatment variable (0 or 1)}

\item{alpha}{tuning parameter for the elastic net}

\item{nfolds}{number of folds for cross-fitting}

\item{lambda.choice}{how to cross-validate}

\item{rs}{whether to use the RS-learner (logical)}

\item{w.hat}{user-supplied estimate for E[W|X]}

\item{y.hat}{user-supplied estimte for E[Y|X]}
}
\description{
R-learner, as proposed by Nie and Wager, 2017
}
