% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_meta_learner_tests}
\alias{simple_meta_learner_tests}
\title{helper function for testing the code runs}
\usage{
simple_meta_learner_tests(tau_hat, sim_data, mse = 0.01)
}
\arguments{
\item{tau_hat}{user-supplied treatment effect estimate}

\item{sim_data}{simulated groundtruth data}

\item{mse}{user-supplied error tolerance}
}
\description{
helper function to check if the learned tau_hat is numeric
}
