% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{t_toy_data_simulation}
\alias{t_toy_data_simulation}
\title{Toy data simulation for T-learner}
\usage{
t_toy_data_simulation(n)
}
\arguments{
\item{n}{the number of samples to draw from the distribution}
}
\value{
a list containing the covariate matrix, treatment vector, outcome vector, true propensity vector, true marginal outcome vector,
true control potential outcome vector, true treated potential outcome vector, and true treatment effect vector, in that order.
}
\description{
Generates a toy dataset of size \eqn{n} that can be used to experiment with T-learners
in this package. The generative process should be easy to learn with linear methods.
}
\examples{
t_toy_data_simulation # show the code- you can modify it to make your own simulations
library(zeallot) # imports the \%<-\% operator, which is syntactic sugar that performs multiple assignment out of a list
c(x, w, y, ...) \%<-\% t_toy_data_simulation(500) # draw a sample
# see what kind of objects these are
str(x)
str(w)
str(y)
}
