% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkern.R
\name{tkern}
\alias{tkern}
\title{T-learner, implemented via kernel ridge regression with Gaussian kernels}
\usage{
tkern(x, w, y, b_range = 10^(seq(-3, 3, 0.5)),
  lambda_range = 10^(seq(-3, 3, 0.5)), k_folds = NULL)
}
\arguments{
\item{x}{the input features}

\item{w}{the treatment variable (0 or 1)}

\item{y}{the observed response (real valued)}

\item{b_range}{the range of Gaussian kernel bandwidths for cross validation}

\item{lambda_range}{the range of ridge regression penalty factor for cross validation}

\item{k_folds}{number of folds for cross validation}
}
\value{
an tkern object
}
\description{
T-learner learns the treated and control expected outcome respectively by fitting two separate models.
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

tkern_fit = tkern(x, w, y)
tkern_est = predict(tkern_fit, x)
}
}
