% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xboost.R
\name{xboost}
\alias{xboost}
\title{X-learner, as proposed by Künzel, Sekhon, Bickel, and Yu 2017}
\usage{
xboost(X, Y, W, nfolds.1 = NULL, nfolds.0 = NULL, nfolds.W = NULL)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{W}{the treatment variable (0 or 1)}

\item{nfolds.1}{number of folds for learning E[Y|X,W=1]}

\item{nfolds.0}{number of folds for learning E[Y|X,W=0]}

\item{nfolds.W}{number of folds for learning E[W|X]}
}
\description{
X-learner, as proposed by Künzel, Sekhon, Bickel, and Yu 2017
}
