% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlasso.R
\name{xlasso}
\alias{xlasso}
\title{X-learner, as proposed by Künzel, Sekhon, Bickel, and Yu 2017}
\usage{
xlasso(X, Y, W, alpha = 1, nfolds.1 = NULL, nfolds.0 = NULL,
  nfolds.W = NULL, lambda.choice = c("lambda.min", "lambda.1se"))
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{W}{the treatment variable (0 or 1)}

\item{alpha}{tuning parameter for the elastic net}

\item{nfolds.1}{number of folds for learning E[Y|X,W=1]}

\item{nfolds.0}{number of folds for learning E[Y|X,W=0]}

\item{nfolds.W}{number of folds for learning E[W|X]}

\item{lambda.choice}{how to cross-validate}
}
\description{
X-learner, as proposed by Künzel, Sekhon, Bickel, and Yu 2017
}
