% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlearner.R
\name{xlearner_cv}
\alias{xlearner_cv}
\title{X-learner using generic black-box machine learning model from caret}
\usage{
xlearner_cv(x, w, y, tau_model_specs, p_model_specs = tau_model_specs,
  mu_model_specs = tau_model_specs, mu0_hat = NULL, mu1_hat = NULL,
  k_folds = 5, select_by = "best", p_min = 0.05, p_max = 0.95)
}
\arguments{
\item{x}{a numeric matrix of \strong{covariates}}

\item{w}{a logical vector indicating \strong{treatment}}

\item{y}{a numeric vector of \strong{outcomes}}

\item{tau_model_specs}{specification for the model of \eqn{\tau(x) = E[Y(1) - Y(0)|X=x]}. See \code{\link{learner_cv}}.}

\item{p_model_specs}{specification for the model of \eqn{p(x) = E[W|X=x]}. See \code{\link{learner_cv}}.
Not needed if \code{p_hat} is provided.}

\item{mu_model_specs}{specification for the models of \eqn{m_w(x) = E[Y|W=w,X=x]}. See \code{\link{learner_cv}}.
Not needed if \code{mu0_hat_0} and \code{mu0_hat} are provided.}

\item{mu0_hat}{a numeric vector of estimates of the outcome under control for all observations.
The X-learner will estimate these values internally if not provided.}

\item{mu1_hat}{a numeric vector of estimates of the outcome under treatment for all observations.
The X-learner will estimate these values internally if not provided.}

\item{k_folds}{number of cross-validation folds to use in hyperparameter optimization for each model.}

\item{select_by}{optimization method to use for cross-validation in each model: either \code{"best"} for minimum cross-validation
error or \code{"oneSE"} for the one-standard-error (1-SE) rule. The implementaion of the 1-SE rule for learners with
multiple hyperparameters is governed by \pkg{caret} and may be ad-hoc for some learners. See: \code{\link[caret]{?caret::oneSE}}.}

\item{p_min}{If provided, estimated propensities will be trimmed to have minimum \code{p_min}.}

\item{p_max}{If provided, estimated propensities will be trimmed to have maximum \code{p_max}.}
}
\description{
X-learner as proposed by Kunzel, Sekhon, Bickel, and Yu (2017), using generic black-box machine learning model from caret
}
\examples{
\dontrun{
model_specs = list(
gbm = list(
    tune_grid = expand.grid(
        n.trees = seq(1,501,20),
        interaction.depth=3,
        shrinkage = 0.1,
        n.minobsinnode=3),
    extra_args = list(
        verbose=F,
        bag.fraction=1)),
glmnet = list(
    tune_grid = expand.grid(
       alpha=c(0,0.5,1),
       lambda=exp(seq(-5,2,0.2))),
    extra_args = list())
)
library(zeallot) # imports the \%<-\% operator, which is syntactic sugar that performs multiple assignment out of a list
c(x, w, y, ...) \%<-\% toy_data_simulation(500) # draw a sample

tau_hat_model = xlearner_cv(x, w, y, model_specs)
tau_hat = predict(tau_hat_model, x)
}
}
