% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvboost.R
\name{cvboost}
\alias{cvboost}
\title{Title}
\usage{
cvboost(X, Y, weights = NULL, nfolds = NULL, objective = c("reg:linear",
  "binary:logistic"), ntrees.max = 30000, num.search.rounds = 20,
  print.every.n = 100, early.stopping.rounds = 10, bayes.opt = FALSE)
}
\arguments{
\item{bayes.opt}{}
}
\description{
Title
}
