% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_simulation}
\alias{data_simulation}
\title{data simulation}
\usage{
data_simulation(n)
}
\arguments{
\item{n}{the number of samples to draw from the distribution}
}
\value{
a list containing the covariate matrix, treatment vector, outcome vector, true propensity vector, true marginal outcome vector,
true control potential outcome vector, true treated potential outcome vector, and true treatment effect vector, in that order.
}
\description{
Generates a dataset of size \eqn{n} that can be used to experiment with the learners and meta-learners
in this package.
}
\examples{
data = data_simulation(500) # draw a sample
}
