% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slearner.R
\name{predict.slearner}
\alias{predict.slearner}
\title{Prediction for S-learner}
\usage{
\method{predict}{slearner}(object, newx, ...)
}
\arguments{
\item{object}{a S-learner object}

\item{newx}{a matrix of covariates for which to predict the treatment effect}
}
\examples{
\dontrun{
model_specs = list(
gbm = list(
    tune_grid = expand.grid(
        n.trees = seq(1,501,20),
        interaction.depth=3,
        shrinkage = 0.1,
        n.minobsinnode=3),
    extra_args = list(
        verbose=F,
        bag.fraction=1)),
glmnet = list(
    tune_grid = expand.grid(
       alpha=c(0,0.5,1),
       lambda=exp(seq(-5,2,0.2))),
    extra_args = list())
)
library(zeallot) # imports the \%<-\% operator, which is syntactic sugar that performs multiple assignment out of a list
c(x, w, y, ...) \%<-\% toy_data_simulation(500) # draw a sample

tau_hat_model = slearner_cv(x, w, y, model_specs)
tau_hat = predict(tau_hat_model, x)
}
}
