% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlasso.R
\name{predict.tlasso}
\alias{predict.tlasso}
\title{predict for tlasso}
\usage{
\method{predict}{tlasso}(object, newx = NULL, s = c("lambda.min",
  "lambda.1se"), ...)
}
\arguments{
\item{object}{a tlasso object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{s}{choose from "lambda.min" or "lambda.1se" for prediction}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained tlasso model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0_5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

tlasso_fit = tlasso(x, w, y)
tlasso_est = predict(tlasso_fit, x)
}


}
