% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulasso.R
\name{predict.ulasso}
\alias{predict.ulasso}
\title{predict for ulasso}
\usage{
\method{predict}{ulasso}(object, newx = NULL, s = c("lambda.1se", "lambda.min"), ...)
}
\arguments{
\item{object}{a ulasso object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{s}{choose from "lambda.min" or "lambda.1se" for prediction}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained ulasso model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

ulasso_fit = ulasso(x, w, y)
ulasso_est = predict(ulasso_fit, x)
}


}
