% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkern.R
\name{predict.xkern}
\alias{predict.xkern}
\title{predict for xkern}
\usage{
\method{predict}{xkern}(object, newx = NULL, new_p_hat = NULL, ...)
}
\arguments{
\item{object}{an xkern object}

\item{newx}{covariate matrix to make predictions on. If null, return the tau(x) predictions on the training data}

\item{new_p_hat}{propensity score on newx provided by the user. Default to NULL. If the user provided their own propensity p_hat in training, new_p_hat must be provided here.}

\item{...}{additional arguments (currently not used)}
}
\value{
vector of predictions
}
\description{
get estimated tau(x) using the trained xkern model
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

xkern_fit = xkern(x, w, y)
xkern_est = predict(xkern_fit, x)
}


}
