% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlasso.R
\name{rlasso}
\alias{rlasso}
\title{R-learner, implemented via glmnet (lasso)}
\usage{
rlasso(
  x,
  w,
  y,
  alpha = 1,
  k_folds = NULL,
  foldid = NULL,
  lambda_y = NULL,
  lambda_w = NULL,
  lambda_tau = NULL,
  lambda_choice = c("lambda.min", "lambda.1se"),
  rs = FALSE,
  p_hat = NULL,
  m_hat = NULL,
  penalty_factor = NULL
)
}
\arguments{
\item{x}{the input features}

\item{w}{the treatment variable (0 or 1)}

\item{y}{the observed response (real valued)}

\item{alpha}{tuning parameter for the elastic net}

\item{k_folds}{number of folds for cross-fitting}

\item{foldid}{user-supplied foldid. Must have length equal to length(w). If provided, it overrides the k_folds option.}

\item{lambda_y}{user-supplied lambda sequence for cross validation in learning E[y|x]}

\item{lambda_w}{user-supplied lambda sequence for cross validation in learning E[w|x]}

\item{lambda_tau}{user-supplied lambda sequence for cross validation in learning the treatment effect E[y(1) - y(0) | x]}

\item{lambda_choice}{how to cross-validate for learning the treatment effect tau; choose from "lambda.min" or "lambda.1se"}

\item{rs}{whether to use the RS-learner (logical).}

\item{p_hat}{user-supplied estimate for E[W|X]}

\item{m_hat}{user-supplied estimte for E[Y|X]}

\item{penalty_factor}{user-supplied penalty factor, a vector of length the same as the number of covariates in x.}
}
\value{
an rlasso object
}
\description{
R-learner, as proposed by Nie and Wager (2017), implemented via glmnet (lasso)
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

rlasso_fit = rlasso(x, w, y)
rlasso_est = predict(rlasso_fit, x)
}
}
