% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sboost.R
\name{sboost}
\alias{sboost}
\title{S-learner, as proposed by Imai and Ratkovic 2013}
\usage{
sboost(X, Y, W, nfolds = NULL)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{W}{the treatment variable (0 or 1)}

\item{nfolds}{number of folds for cross-fitting}
}
\description{
S-learner, as proposed by Imai and Ratkovic 2013
}
