% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slasso.R
\name{slasso}
\alias{slasso}
\title{S-learner, implemented via glmnet (lasso)}
\usage{
slasso(
  x,
  w,
  y,
  alpha = 1,
  k_folds = NULL,
  foldid = NULL,
  lambda = NULL,
  lambda_choice = c("lambda.min", "lambda.1se"),
  penalty_factor = NULL
)
}
\arguments{
\item{x}{the input features}

\item{w}{the treatment variable (0 or 1)}

\item{y}{the observed response (real valued)}

\item{alpha}{tuning parameter for the elastic net}

\item{k_folds}{number of folds for cross validation}

\item{lambda}{user-supplied lambda sequence for cross validation}

\item{lambda_choice}{how to cross-validate; choose from "lambda.min" or "lambda.1se"}

\item{penalty_factor}{user-supplied penalty factor, must be of length the same as number of features in x}
}
\description{
S-learner, as proposed by Imai and Ratkovic (2013), implemented via glmnet (lasso)
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

slasso_fit = slasso(x, w, y)
slasso_est = predict(slasso_fit, x)
}
}
