% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slasso.R
\name{slasso}
\alias{slasso}
\title{S-learner, as proposed by Imai and Ratkovic 2013}
\usage{
slasso(X, Y, W, alpha = 1, nfolds = NULL, lambda.choice = c("lambda.min",
  "lambda.1se"), penalty.search = FALSE)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{W}{the treatment variable (0 or 1)}

\item{alpha}{tuning parameter for the elastic net}

\item{nfolds}{number of folds for cross-fitting}

\item{lambda.choice}{how to cross-validate}

\item{penalty.search}{whether to perform fine grainted penalty factor search (logical)}
}
\description{
S-learner, as proposed by Imai and Ratkovic 2013
}
