% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tboost.R
\name{tboost}
\alias{tboost}
\title{T-learner, implemented via xgboost (gradient boosting)}
\usage{
tboost(x, w, y, alpha = 1, k_folds_mu1 = NULL, k_folds_mu0 = NULL,
  ntrees_max = 1000, num_search_rounds = 10, print_every_n = 100,
  early_stopping_rounds = 10, nthread = NULL, bayes_opt = FALSE)
}
\arguments{
\item{x}{the input features}

\item{w}{the treatment variable (0 or 1)}

\item{y}{the observed response (real valued)}

\item{k_folds_mu1}{number of folds for cross validation for the treated}

\item{k_folds_mu0}{number of folds for cross validation for the control}

\item{ntrees_max}{the maximum number of trees to grow for xgboost}

\item{num_search_rounds}{the number of random sampling of hyperparameter combinations for cross validating on xgboost trees}

\item{print_every_n}{the number of iterations (in each iteration, a tree is grown) by which the code prints out information}

\item{early_stopping_rounds}{the number of rounds the test error stops decreasing by which the cross validation in finding the optimal number of trees stops}

\item{nthread}{the number of threads to use. The default is NULL, which uses all available threads}

\item{bayes_opt}{if set to TRUE, use bayesian optimization to do hyper-parameter search in xgboost. if set to FALSE, randomly draw combinations of hyperparameters to search from (as specified by num_search_rounds). Default is FALSE.}
}
\description{
T-learner, implemented via xgboost (gradient boosting)
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

tboost_fit = tboost(x, w, y)
tboost_est = predict(tboost_fit, x)
}

}
