% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tboost.R
\name{tboost}
\alias{tboost}
\title{Title}
\usage{
tboost(X, Y, W, alpha = 1, nfolds.1 = NULL, nfolds.0 = NULL,
  lambda.choice = c("lambda.min", "lambda.1se"))
}
\arguments{
\item{lambda.choice}{}
}
\description{
Title
}
