% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlasso.R
\name{tlasso}
\alias{tlasso}
\title{T-learner, implemented via glmnet (lasso)}
\usage{
tlasso(x, w, y, alpha = 1, k_folds_mu1 = NULL, k_folds_mu0 = NULL,
  lambda_choice = c("lambda.min", "lambda.1se"))
}
\arguments{
\item{x}{the input features}

\item{w}{the treatment variable (0 or 1)}

\item{y}{the observed response (real valued)}

\item{alpha}{tuning parameter for the elastic net}

\item{k_folds_mu1}{number of folds for cross validation for the treated}

\item{k_folds_mu0}{number of folds for cross validation for the control}

\item{lambda_choice}{how to cross-validate; choose from "lambda.min" or "lambda.1se"}
}
\description{
T-learner, implemented via glmnet (lasso)
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0_5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

tlasso_fit = tlasso(x, w, y)
tlasso_est = predict(tlasso_fit, x)
}
}
