% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uboost.R
\name{uboost}
\alias{uboost}
\title{U-learner implemented via xgboost (boosting)}
\usage{
uboost(
  x,
  w,
  y,
  k_folds = NULL,
  p_hat = NULL,
  m_hat = NULL,
  cutoff = 0.05,
  ntrees_max = 1000,
  num_search_rounds = 10,
  print_every_n = 100,
  early_stopping_rounds = 10,
  nthread = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{the input features}

\item{w}{the treatment variable (0 or 1)}

\item{y}{the observed response (real valued)}

\item{k_folds}{number of folds used for cross fitting and cross validation}

\item{p_hat}{pre-computed estimates on E[W|X] corresponding to the input x. uboost will compute it internally if not provided.}

\item{m_hat}{pre-computed estimates on E[Y|X] corresponding to the input x. uboost will compute it internally if not provided.}

\item{cutoff}{the threshold to cutoff propensity estimate}

\item{ntrees_max}{the maximum number of trees to grow for xgboost}

\item{num_search_rounds}{the number of random sampling of hyperparameter combinations for cross validating on xgboost trees}

\item{print_every_n}{the number of iterations (in each iteration, a tree is grown) by which the code prints out information}

\item{early_stopping_rounds}{the number of rounds the test error stops decreasing by which the cross validation in finding the optimal number of trees stops}

\item{nthread}{the number of threads to use. The default is NULL, which uses all available threads}

\item{verbose}{boolean; whether to print statistic}
}
\description{
U-learner as proposed by Kunzel, Sekhon, Bickel, and Yu (2017), implemented via xgboost (boosting)
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

uboost_fit = uboost(x, y, w)
uboost_est = predict(uboost_fit, x)
}

}
