% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulasso.R
\name{ulasso}
\alias{ulasso}
\title{U-learner}
\usage{
ulasso(X, Y, W, alpha = 1, nfolds = NULL, lambda.choice = c("lambda.min",
  "lambda.1se"), cutoff = 0.05)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{W}{the treatment variable (0 or 1)}

\item{alpha}{tuning parameter for the elastic net}

\item{nfolds}{number of folds for cross-fitting}

\item{lambda.choice}{how to cross-validate}

\item{cutoff}{propensity cutoff threshold}
}
\description{
U-learner
}
