% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulasso.R
\name{ulasso}
\alias{ulasso}
\title{U-learner implemented via glmnet (lasso)}
\usage{
ulasso(x, w, y, alpha = 1, k_folds = NULL, lambda_choice = c("lambda.1se",
  "lambda.min"), p_hat = NULL, m_hat = NULL, cutoff = 0.05)
}
\arguments{
\item{x}{the input features}

\item{w}{the treatment variable (0 or 1)}

\item{y}{the observed response (real valued)}

\item{alpha}{tuning parameter for the elastic net}

\item{k_folds}{number of folds for cross-fitting}

\item{lambda_choice}{how to cross-validate; choose from "lambda.1se" or "lambda.min"}

\item{p_hat}{user-supplied estimate for E[W|X]}

\item{m_hat}{user-supplied estimte for E[Y|X]}

\item{cutoff}{the threshold to cutoff propensity estimate}
}
\description{
U-learner as proposed by Künzel, Sekhon, Bickel, and Yu (2017), implemented via glmnet (lasso)
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

ulasso_fit = ulasso(x, w, y)
ulasso_est = predict(ulasso_fit, x)
}
}
