% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xboost.R
\name{xboost}
\alias{xboost}
\title{X-learner implemented via xgboost (boosting)}
\usage{
xboost(x, w, y, k_folds_mu1 = NULL, k_folds_mu0 = NULL, k_folds_p = NULL,
  mu1_hat = NULL, mu0_hat = NULL, p_hat = NULL, ntrees_max = 1000,
  num_search_rounds = 10, print_every_n = 100, early_stopping_rounds = 10,
  nthread = NULL, verbose = FALSE, bayes_opt = FALSE)
}
\arguments{
\item{x}{the input features}

\item{w}{the treatment variable (0 or 1)}

\item{y}{the observed response (real valued)}

\item{k_folds_mu1}{number of folds for learning E[Y|X,W=1]}

\item{k_folds_mu0}{number of folds for learning E[Y|X,W=0]}

\item{k_folds_p}{number of folds for learning E[W|X]}

\item{mu1_hat}{pre-computed estimates on E[Y|X,W=1] corresponding to the input x. xboost will compute it internally if not provided}

\item{mu0_hat}{pre-computed estimates on E[Y|X,W=0] corresponding to the input x. xboost will compute it internally if not provided}

\item{p_hat}{pre-computed estimates on E[W|X] corresponding to the input x. xboost will compute it internally if not provided}

\item{ntrees_max}{the maximum number of trees to grow for xgboost}

\item{num_search_rounds}{the number of random sampling of hyperparameter combinations for cross validating on xgboost trees}

\item{print_every_n}{the number of iterations (in each iteration, a tree is grown) by which the code prints out information}

\item{early_stopping_rounds}{the number of rounds the test error stops decreasing by which the cross validation in finding the optimal number of trees stops}

\item{nthread}{the number of threads to use. The default is NULL, which uses all available threads}

\item{verbose}{boolean; whether to print statistic}

\item{bayes_opt}{if set to TRUE, use bayesian optimization to do hyper-parameter search in xgboost. if set to FALSE, randomly draw combinations of hyperparameters to search from (as specified by num_search_rounds). Default is FALSE. CAUTION: current implementation is in beta and is not recommended for usage yet}
}
\description{
X-learner as proposed by Künzel, Sekhon, Bickel, and Yu (2017), implemented via xgboost (boosting)
}
\examples{
\dontrun{
n = 100; p = 10

x = matrix(rnorm(n*p), n, p)
w = rbinom(n, 1, 0.5)
y = pmax(x[,1], 0) * w + x[,2] + pmin(x[,3], 0) + rnorm(n)

xboost_fit = xboost(x, w, y)
xboost_est = predict(xboost_fit, x)
}

}
