% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlasso.R
\name{xlasso}
\alias{xlasso}
\title{X-learner, as proposed by Künzel, Sekhon, Bickel, and Yu 2017, implemented via glmnet (lasso)}
\usage{
xlasso(X, Y, W, alpha = 1, nfolds.1 = NULL, nfolds.0 = NULL,
  nfolds.W = NULL, lambda.choice = c("lambda.min", "lambda.1se"),
  y.1.pred = NULL, y.0.pred = NULL)
}
\arguments{
\item{X}{the input features}

\item{Y}{the observed response (real valued)}

\item{W}{the treatment variable (0 or 1)}

\item{alpha}{tuning parameter for the elastic net}

\item{nfolds.1}{number of folds for learning E[Y|X,W=1]}

\item{nfolds.0}{number of folds for learning E[Y|X,W=0]}

\item{nfolds.W}{number of folds for learning E[W|X]}

\item{lambda.choice}{how to cross-validate; choose from "lambda.min" or "lambda.1se"}

\item{y.1.pred}{pre-computed estimates on E[Y|X,W=1] corresponding to the input X. xlasso will compute it internally if not provided.}

\item{y.0.pred}{pre-computed estimates on E[Y|X,W=0] corresponding to the input X. xlasso will compute it internally if not provided.}
}
\description{
X-learner, as proposed by Künzel, Sekhon, Bickel, and Yu 2017, implemented via glmnet (lasso)
}
\examples{
\dontrun{
n = 100; p = 10

X = matrix(rnorm(n*p), n, p)
W = rbinom(n, 1, 0.5)
Y = pmax(X[,1], 0) * W + X[,2] + pmin(X[,3], 0) + rnorm(n)

xlasso.fit = xlasso(X, Y, W)
xlasso.est = predict(xlasso.fit, X)
}


}
