% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_utils.R
\name{xval_xfit}
\alias{xval_xfit}
\title{Cross-validated cross-fitting}
\usage{
xval_xfit(x, y, model_specs, economy = T, weights = NULL,
  k_folds_cf = 5, k_folds = 5, select_by = "best")
}
\arguments{
\item{x}{a numeric matrix of features}

\item{y}{a logical vector for probabilistic classificaton or numeric vector for regression.}

\item{model_specs}{a data structure specifying which learning algorithms, hyperparameters should
be cross validated over, and which additional arguments should be
passed to each learner. This should be a list where the names of each element are valid \pkg{caret}
methods (learning algorithms). The list element corresponding to each learning algorithm should itself
be a list of two elements named \code{tune_grid} (hyperparameters) and \code{extra_args}.
\code{tune_grid} should be a valid \pkg{caret} tune grid of hyperparameters corresponing to the learning algorithm.
\code{extra_args} is a named list of additional arguments to be passed on to the learning algorithm. See example.}

\item{economy}{flag that determines if "economy" or "deluxe" cross-validated cross-estimation is performed}

\item{weights}{optional case weights. Note that only some caret models support this option for weighted regression. Details see https://topepo.github.io/caret/train-models-by-tag.html#accepts-case-weights}

\item{k_folds_cf}{number of cross-fitting folds. Unecessary if \code{economy=T}.}

\item{k_folds}{number of cross-validation folds}

\item{select_by}{optimization method to use for cross-validation: either \code{"best"} for minimum cross-validation
error or \code{"oneSE"} for the one-standard-error (1-SE) rule. The implementaion of the 1-SE rule for learners with
multiple hyperparameters is governed by \pkg{caret} and may be ad-hoc for some learners. See: \code{\link[caret]{?caret::oneSE}}.}
}
\description{
Cross-validated cross-fitting
}
\details{
Provides both a "deluxe" version and an "economy" version. The deluxe version
preserves data-splitting independence relations. The "economy version" leaks information from the held-out folds
into the predictions on the held-out folds via the hyperparameter selection: data-splitting independence assumptions
do not hold and theoretical guarentees do not follow, but the models fit more quickly. Internal cross validation is
performed via \code{\link{learner_cv}}.
}
\examples{
\dontrun{
model_specs = list(
gbm = list(
    tune_grid = expand.grid(
        n.trees = seq(1,501,20),
        interaction.depth=3,
        shrinkage = 0.1,
        n.minobsinnode=3),
    extra_args = list(
        verbose=F,
        bag.fraction=1)),
glmnet = list(
    tune_grid = expand.grid(
       alpha=c(0,0.5,1),
       lambda=exp(seq(-5,2,0.2))),
    extra_args = list())
)
library(zeallot) # imports the \%<-\% operator, which is syntactic sugar that performs multiple assignment out of a list
c(x, w, y, ...) \%<-\% toy_data_simulation(500) # draw a sample

y_hat = xval_xfit(x, y, model_specs)
w_hat_prob = xval_xfit(x, w, model_specs)
}
}
